revise the UpdateCourse method to make it also able to create new post semester depending on the year inputted.
